+++ pkg/resource/deploy/source_error.go
@@ -35,6 +35,7 @@ type errorSource struct {
 
 func (src *errorSource) Close() error                { return nil }
 func (src *errorSource) Project() tokens.PackageName { return src.project }
+func (src *errorSource) Info() interface{}           { return nil }
 
 func (src *errorSource) Iterate(
 	ctx context.Context, opts Options, providers ProviderSource) (SourceIterator, result.Result) {