+++ scripts/build-docker.sh
@@ -87,7 +87,7 @@ publish_containers() {
     docker logout
 
     # This publishes the SDK specific containers and uses a dispatch event to trigger a GitHub Action
-    pulumictl create containers ${CLI_VERSION}
+    pulumictl create containers "${CLI_VERSION//v}"
 }
 
 echo_header "Building Pulumi containers (${CLI_VERSION})"