+++ sdk/go/common/resource/config/value.go
@@ -72,6 +72,46 @@ func (c Value) Value(decrypter Decrypter) (string, error) {
 	return decrypter.DecryptValue(c.value)
 }
 
+func (c Value) Copy(decrypter Decrypter, encrypter Encrypter) (Value, error) {
+	var val Value
+	raw, err := c.Value(decrypter)
+	if err != nil {
+		return Value{}, err
+	}
+	if c.Secure() {
+		if c.Object() {
+			objVal, err := c.ToObject()
+			if err != nil {
+				return Value{}, err
+			}
+			encryptedObj, err := reencryptObject(objVal, decrypter, encrypter)
+			if err != nil {
+				return Value{}, err
+			}
+			json, err := json.Marshal(encryptedObj)
+			if err != nil {
+				return Value{}, err
+			}
+
+			val = NewSecureObjectValue(string(json))
+		} else {
+			enc, eerr := encrypter.EncryptValue(raw)
+			if eerr != nil {
+				return Value{}, eerr
+			}
+			val = NewSecureValue(enc)
+		}
+	} else {
+		if c.Object() {
+			val = NewObjectValue(raw)
+		} else {
+			val = NewValue(raw)
+		}
+	}
+
+	return val, nil
+}
+
 func (c Value) SecureValues(decrypter Decrypter) ([]string, error) {
 	d := NewTrackingDecrypter(decrypter)
 	if _, err := c.Value(d); err != nil {
@@ -238,6 +278,53 @@ func isSecureValue(v interface{}) (bool, string) {
 		}
 	}
 	return false, ""
+}
+
+func reencryptObject(v interface{}, decrypter Decrypter, encrypter Encrypter) (interface{}, error) {
+	reencryptIt := func(val interface{}) (interface{}, error) {
+		if isSecure, secureVal := isSecureValue(val); isSecure {
+			newVal := NewSecureValue(secureVal)
+			raw, err := newVal.Value(decrypter)
+			if err != nil {
+				return nil, err
+			}
+
+			encVal, err := encrypter.EncryptValue(raw)
+			if err != nil {
+				return nil, err
+			}
+
+			m := make(map[string]string)
+			m["secure"] = encVal
+
+			return m, nil
+		}
+		return reencryptObject(val, decrypter, encrypter)
+	}
+
+	switch t := v.(type) {
+	case map[string]interface{}:
+		m := make(map[string]interface{})
+		for key, val := range t {
+			encrypted, err := reencryptIt(val)
+			if err != nil {
+				return nil, err
+			}
+			m[key] = encrypted
+		}
+		return m, nil
+	case []interface{}:
+		a := make([]interface{}, len(t))
+		for i, val := range t {
+			encrypted, err := reencryptIt(val)
+			if err != nil {
+				return nil, err
+			}
+			a[i] = encrypted
+		}
+		return a, nil
+	}
+	return v, nil
 }
 
 // decryptObject returns a new object with all secure values in the object converted to decrypted strings.