+++ sdk/python/lib/pulumi/resource.py
@@ -23,15 +23,11 @@ from .runtime.settings import get_root_resource
 
 from .metadata import get_project, get_stack
 
-from .output import Output
-
 if TYPE_CHECKING:
-    from .output import Input, Inputs
+    from .output import Input, Inputs, Output
     from .runtime.stack import Stack
 
 
-
-@known_types.custom_timeouts
 class CustomTimeouts:
     create: Optional[str]
     """
@@ -81,6 +77,7 @@ def inherited_child_alias(
 #   * parentAliasName: "app"
 #   * aliasName: "app-function"
 #   * childAlias: "urn:pulumi:stackname::projectname::aws:s3/bucket:Bucket::app-function"
+    from . import Output  # pylint: disable=import-outside-toplevel
     alias_name = Output.from_input(child_name)
     if child_name.startswith(parent_name):
         alias_name = Output.from_input(parent_alias).apply(
@@ -163,14 +160,14 @@ class Alias:
     """
 
     # Ignoring type errors associated with the ellipsis constant being assigned to a string value.
-    # We use it as a internal sentinal value, and don't need to expose this in the user facing type system.
+    # We use it as a internal sentinel value, and don't need to expose this in the user facing type system.
     # https://docs.python.org/3/library/constants.html#Ellipsis
     def __init__(self,
-                 name: Optional[str] = ..., # type: ignore
-                 type_: Optional[str] = ..., # type: ignore
-                 parent: Optional[Union['Resource', 'Input[str]']] = ..., # type: ignore
-                 stack: Optional['Input[str]'] = ..., # type: ignore
-                 project: Optional['Input[str]'] = ...) -> None: # type: ignore
+                 name: Optional[str] = ...,  # type: ignore
+                 type_: Optional[str] = ...,  # type: ignore
+                 parent: Optional[Union['Resource', 'Input[str]']] = ...,  # type: ignore
+                 stack: Optional['Input[str]'] = ...,  # type: ignore
+                 project: Optional['Input[str]'] = ...) -> None:  # type: ignore
 
         self.name = name
         self.type_ = type_
@@ -187,8 +184,9 @@ def collapse_alias_to_urn(
     """
     collapse_alias_to_urn turns an Alias into a URN given a set of default data
     """
+    from . import Output  # pylint: disable=import-outside-toplevel
 
-    def collapse_alias_to_urn_worker(inner: Union[Alias, str]) -> 'Output[str]':
+    def collapse_alias_to_urn_worker(inner: Union[Alias, str]) -> Output[str]:
         if isinstance(inner, str):
             return Output.from_input(inner)
 
@@ -208,6 +206,7 @@ def collapse_alias_to_urn(
 
     inputAlias: Output[Union[Alias, str]] = Output.from_input(alias)
     return inputAlias.apply(collapse_alias_to_urn_worker)
+
 
 class ResourceTransformationArgs:
     """
@@ -251,6 +250,7 @@ class ResourceTransformationArgs:
         self.props = props
         self.opts = opts
 
+
 class ResourceTransformationResult:
     """
     ResourceTransformationResult is the result that must be returned by a resource transformation
@@ -274,6 +274,7 @@ class ResourceTransformationResult:
         self.props = props
         self.opts = opts
 
+
 ResourceTransformation = Callable[[ResourceTransformationArgs], Optional[ResourceTransformationResult]]
 """
 ResourceTransformation is the callback signature for the `transformations` resource option.  A
@@ -283,6 +284,7 @@ actually being created.  The effect will be as though those props and opts were 
 of the original call to the `Resource` constructor.  If the transformation returns undefined,
 this indicates that the resource will not be transformed.
 """
+
 
 class ResourceOptions:
     """
@@ -458,7 +460,7 @@ class ResourceOptions:
             values from each options object. Both original collections in each options object will
             be unchanged.
 
-        2. Simple scaler values from `opts2` (i.e. strings, numbers, bools) will replace the values
+        2. Simple scalar values from `opts2` (i.e. strings, numbers, bools) will replace the values
             from `opts1`.
 
         3. For the purposes of merging `depends_on`, `provider` and `providers` are always treated
@@ -551,6 +553,7 @@ def _merge_lists(dest, source):
 
     return dest + source
 
+
 # !!! IMPORTANT !!! If you add a new attribute to this type, make sure to verify that merge_options
 # works properly for it.
 class Resource:
@@ -621,8 +624,8 @@ class Resource:
         elif not isinstance(opts, ResourceOptions):
             raise TypeError('Expected resource options to be a ResourceOptions instance')
 
-        # Before anything else - if there are transformations registered, give them a chance to run to modify the user provided
-        # properties and options assigned to this resource.
+        # Before anything else - if there are transformations registered, give them a chance to run to modify the user
+        # provided properties and options assigned to this resource.
         parent = opts.parent
         if parent is None:
             parent = get_root_resource()
@@ -667,7 +670,7 @@ class Resource:
             for parent_alias in opts.parent._aliases:
                 child_alias = inherited_child_alias(
                     name, opts.parent._name, parent_alias, t)
-                opts.aliases.append(cast(Output[Union[str, Alias]], child_alias))
+                opts.aliases.append(cast('Output[Union[str, Alias]]', child_alias))
 
             # Infer providers and provider maps from parent, if one was provided.
             self._providers = opts.parent._providers
@@ -776,7 +779,6 @@ class Resource:
         return self._providers.get(pkg)
 
 
-@known_types.custom_resource
 class CustomResource(Resource):
     """
     CustomResource is a resource whose create, read, update, and delete (CRUD) operations are
@@ -903,7 +905,7 @@ def create_urn(
     create_urn computes a URN from the combination of a resource name, resource type, optional
     parent, optional project and optional stack.
     """
-
+    from . import Output  # pylint: disable=import-outside-toplevel
     parent_prefix: Optional[Output[str]] = None
     if parent is not None:
         parent_urn = None